refactor(server): split 2847-line server.js into route modules (#75)
Extract all HTTP route handlers from the monolithic server.js into
11 focused modules under server/routes/, reducing server.js from
3073 to 213 lines. Pure refactor with zero behavior changes.

Route modules:
- _shared.js: common helpers (conversationById, pushMessage, etc.)
- push.js: POST/DELETE /api/push-token
- vault.js: /api/vault/* CRUD
- usage.js: /api/usage, /api/usage/summary
- controls.js: GET/POST /api/controls
- github.js: /api/github/* proxy routes
- evolution.js: /api/signals, /api/insights, /api/lessons
- briefs.js: /api/brief/*, /brief/* UI
- chat.js: /api/conversations, /api/participants + processQueue
- jira.js: /api/webhooks/jira, /api/integrations/jira
- tasks.js: /api/tasks/*, /api/dispatch-next, /api/project, /api/retro

Each module exports: function(req, res, helpers, deps) => boolean
Dependencies injected via deps object assembled in server.js.
